refactor(server): split mcpServer.ts into core + registries + legacy + capabilities + resourceTemplate
- mcpServer.ts: 1795 -> 933 LOC (core class, handle/connect/buildContext, server->client methods)
- serverRegistries.ts: 873 (registerTool/Resource/Prompt + lazy installers, composed via _registries)
- serverCapabilities.ts: 130 (assert* functions, method->capability table)
- serverLegacy.ts: 105 (v1 overload parsing helpers)
- resourceTemplate.ts: 45

All tests/lint/typecheck green.
